Arreglo CPP-VALIDATORS y pseudos
Ahora si, he separado la definicion de la clase de las definiciones, correspondientes a los metodos de clase, en cabecera y cpp respectivamente,

Tema aparte actualice los algoritmos para incluir los algoritmos usando regex. @werfenGranate45
